refactor: move process threshold function
